step 분리
fewer hook err 해결
